Break out some methods
