+++ lib/prx_auth/rails/engine.rb
@@ -4,6 +4,7 @@ module PrxAuth
       config.to_prepare do
         ::ApplicationController.helper_method [
           :current_user, :prx_jwt,
+          :current_user_access?, :current_user_admin?, :current_user_wildcard?,
           :current_user_info, :current_user_name, :current_user_apps,
           :account_name_for, :account_for, :accounts_for
         ]
